issue#207 : Move tests to RDFWriterTest for all writer/parser combinations
The patterns shouldn't fail for any writer/parser combinations.
